fashion_finder_fixed/requirements.txt
flask==2.2.3
flask-cors==3.0.10
flask-jwt-extended==4.4.4
numpy==1.24.2
pandas==1.5.3
scikit-learn==1.2.2
pillow==9.4.0
requests==2.28.2
python-dotenv==1.0.0
scipy==1.10.1
matplotlib==3.7.1
joblib==1.2.0
tqdm==4.65.0
